Add the new IPC and ring buffer files.
Adds file_contexts entries for the files representing new AOC services
used by a new transport between CHRE on AOC and the bthal.

NOTE: The use-case for these files is the same as the
/dev/acd-com.google.bt* files (lines 23-24). Those files will eventually
be removed when we migrate from the old USF-based transport to the one
under development.

Bug: 342415506
Test: Works with other CLs in topic to support BT offload over the new
transport (with a local change to hook the new transport in).

Change-Id: I81246a25af020ab14da3d4ae7675f2a0f00f97be