Limit maxGxpDebugDumps to 4
DSP multicore workload can run atmost on 4 cores(in case of
amalthea) thus in case of crash we will need atmost 4 core
dumps. Thus limiting maxGxpDebugDumps to 4 to make sure that
we will be capturing all the required core dumps for the
recent crash and at the same time not inflating the bugreport
size.

Bug: 338583087
Change-Id: I76c5b1199c0d89f2bd9a2720dc24a7c86d057b4c