gs_watchdogd: Use timeout defined by platform
Uses timeout value of watchdog device directly, without modifying it
with ioctl. On some platforms, especially these with multiple watchdog
devices, the timeout values can be strictly defined for the watchdogs to
timeout in a certain order. Leaves timeout value as it is and divides
minimal of them by 2 as watchdog kicking period.

Flag: EXEMPT normal/day-to-day bugfix
Bug: 348318712
Change-Id: Icdcce368f1803cd3b38a48f05e2788d881e3ad6f